Change the update process for choco packages
In version 1.0.0 of chocolatey the update ability will be
removed, and the new option "upgrade" will no longer install if
the package doesn't exist yet. This change makes a check for the
package on the local system and then install/upgrades appropriately
removing a crap load of warnings and sets this script up to handle
the future version of Chocolatey.